Scripts
Better get configs

+++ scripts/packages/get_configs.sh
@@ -13,47 +13,49 @@ else
 fi
 
 mkdir -p "$confd/nvim"
-stow -t "$confd/nvim" "$dotdir/nvim"
+stow -t "$confd/nvim" "nvim"
 
 mkdir -p "$confd/paru"
-stow -t "$confd/paru" "$dotdir/paru"
+stow -t "$confd/paru" "paru"
 
 mkdir -p "$confd/qtile"
-stow -t "$confd/qtile" "$dotdir/qtile"
+stow -t "$confd/qtile" "qtile"
 
 mkdir -p "$confd/ranger"
-stow -t "$confd/ranger" "$dotdir/ranger"
+stow -t "$confd/ranger" "ranger"
 
 mkdir -p "$confd/X11"
-stow -t "$confd/X11" "$dotdir/X11"
+stow -t "$confd/X11" "X11"
 
 mkdir -p "$confd/xplr"
-stow -t "$confd/xplr" "$dotdir/xplr"
+stow -t "$confd/xplr" "xplr"
 
 mkdir -p "$confd/hypr"
-# stow -t "$confd/hypr" "$dotdir/hyprland"
-stow -t "$confd/hypr" "$dotdir/hypr"
+# stow -t "$confd/hypr" "hyprland"
+stow -t "$confd/hypr" "hypr"
 
 mkdir -p "$confd/yazi"
-stow -t "$confd/yazi" "$dotdir/yazi"
+stow -t "$confd/yazi" "yazi"
 
 mkdir -p "$confd/waybar"
-stow -t "$confd/waybar" "$dotdir/waybar"
+stow -t "$confd/waybar" "waybar"
 
 mkdir -p "$confd/walker"
-stow -t "$confd/walker" "$dotdir/walker"
+stow -t "$confd/walker" "walker"
 
 mkdir -p "$confd/dunst"
-stow -t "$confd/dunst" "$dotdir/dunst"
+stow -t "$confd/dunst" "dunst"
 
 mkdir -p "$confd/wlogout"
-stow -t "$confd/wlogout" "$dotdir/wlogout"
+stow -t "$confd/wlogout" "wlogout"
 
-stow -t "$HOME" "$dotdir/gdu"
-doas cp "$HOME/.dotfiles/gdu/.gdu.yaml" /root
+stow -t "$HOME" "gdu"
+sudo cp "$HOME/.dotfiles/gdu/.gdu.yaml" /root
 
-cp ./git/.gitconfig "$HOME"
-echo "Do not forget to add user section in .gitconfig"
+if [[ ! -f "$HOME/.gitconfig"]]; then
+    cp ./git/.gitconfig "$HOME"
+    echo "Do not forget to add user section in .gitconfig"
+fi
 
 # if [ ! -d "/etc/lemurs/wms" ] || [ ! -d "/etc/lemurs/wayland" ]; then
 echo "Lemurs setup needs password to write files"
@@ -65,20 +67,20 @@ sudo cp -r "$HOME/.dotfiles/lemurs/wayland" "/etc/lemurs/wayland"
 
 echo "Greetd need permission to write files"
 sudo mkdir -p /etc/greetd/
-sudo cp "$HOME/.dotfiles/greetd/config.toml" "/etc/greetd/"
+sudo cp -r "$HOME/.dotfiles/greetd"/* "/etc/greetd/"
 
 echo "Reflector needs password to write files"
 sudo cp "$HOME/.dotfiles/reflector/refletor.conf" "/etc/xdg/reflector/reflector.conf"
 sudo systemctl enable reflector.timer
 sudo systemctl start reflector.timer
 
-stow -t "$HOME" "$dotdir/tmux"
+stow -t "$HOME" "tmux"
 if [ ! -d "$HOME/.tmux/plugins/tpm/.git" ]; then
     git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 fi
 
 # stow -t "$HOME" scripts 
-stow -t "$HOME" "$dotdir/powerlevel10k"
-stow -t "$HOME" "$dotdir/zsh" 
+stow -t "$HOME" "powerlevel10k"
+stow -t "$HOME" "zsh" 
 source ~/.zshrc && . "$SCRSDIR/zsh_plugin&script_downloader.sh"
